requirements.txt
openai==1.88.0
python-dotenv==1.1.0
telegram==0.0.1